fix(ios): resolve MOB.8 SwiftLint file/type length failures
Raise Boards folder lint ceilings for BoardDetailView growth, and split
MOB.8 advanced board models into LMSBoardAdvancedModels.swift so
LMSBoardModels stays under the Core/LMS file_length error limit.
